Fix notification
